+++ packages/client/src/hoks/RequaredAuth.tsx
@@ -1,6 +1,6 @@
 import { Navigate, useLocation } from 'react-router-dom'
-import { useAuth } from '../../hooks/useAuth'
-import { Routes } from './routes'
+import { useAuth } from '../hooks/useAuth'
+import { Routes } from '../utils/routes'
 
 export function RequaredAuth({ children }: { children: JSX.Element }) {
   const { user } = useAuth();
